third_party/blink/web_tests/touchadjustment/event-triggered-widgets-expected.txt

Test various ways to trigger input-widgets. On a touch interface, all the actions should be triggerable with either a touch down or a touch tap.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing small direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
Testing large direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
Testing large direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
PASS successfullyParsed is true

TEST COMPLETE
